% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvmr.pacs.datathin.R
\name{mvmr.pacs.datathin}
\alias{mvmr.pacs.datathin}
\title{Function to perform MVMR-PACS estimator for signal-group identification and variable selection, along with two-fold data thinning for post-selection inference.}
\usage{
mvmr.pacs.datathin(
  beta.exposure,
  se.exposure,
  beta.outcome,
  se.outcome,
  P,
  type = 2,
  rr_cut_off = 0,
  tau = c(0.5, 1, 2, 3),
  fold = 5,
  over.dispersion = FALSE,
  re = 1,
  n_times = 1,
  epsilon = 1e-04,
  digit = 3,
  include_zero_group = FALSE,
  lambda.1se = TRUE
)
}
\arguments{
\item{beta.exposure}{A data.frame or matrix. Each row contains the estimated marginal effect of a SNP on K exposures, usually obtained from a GWAS}

\item{se.exposure}{A data.frame or matrix of estimated standard errors of beta.exposure}

\item{beta.outcome}{A vector of the estimated marginal effect of a SNP on outcome, usually obtained from a GWAS}

\item{se.outcome}{A vector of estimated standard errors of beta.outcome}

\item{P}{A K-by-K matrix for the estimated shared correlation matrix between the effect of the genetic variants on each exposure, where K is the number of exposure.}

\item{type}{Weighting scheme in penalization: 1 = weights independent of correlation between beta.exposure; 2 = weights inversely related to correlation between beta.exposure; 3 = weights non-zero only if absolute correlation between beta.exposure greater than rr_cut_off; 4 = weights inversely related to correlation between beta.exposure only if absolute correlation between beta.exposure greater than rr_cut_off.}

\item{rr_cut_off}{Pre-specified cut-off used in different weighting schemes, specified by type argument.}

\item{tau}{Tuning parameter range. Default = c(0.5, 1, 2, 3).}

\item{fold}{The number of folds for cross validation based on multifold data thinning. Default = 5.}

\item{over.dispersion}{Should SRIVW consider balanced horizontal pleiotropy when performing inference? Default is FALSE.}

\item{re}{The number of repeated data thinning procedure. Default = 1. Increase this number of assess robustness of results.}

\item{n_times}{The number of cross validation runs when selecting tuning parameters within the selection dataset. Default = 1.}

\item{epsilon}{Tolerance level for ADMM.proj. Default = 1e-4.}

\item{digit}{Number of digits to keep when determining signal-groups. Default = 3, i.e., decimal precision to 0.001.}

\item{include_zero_group}{Whether or not consider non-causal-group identified by MVMR-PACS from the selection dataset as a group for post-selection inference. Default = FALSE.}

\item{lambda.1se}{Whether or not to perform 1SE rules when selecting tuning parameters. Default = TRUE.}
}
\value{
A list with elements
\item{grouping.hist}{Identified signal-groups across repeated data thinning runs.}
\item{beta.hist}{Matrix of estimated direct effect of each signal-group on the outcome from SRIVW across repeated data thinning runs.}
\item{se.hist}{Matrix of estimated standard error of direct effect of each signal-group on the outcome from SRIVW across repeated data thinning runs.}
\item{iv.hist.dt1}{Sample IV strength parameter in the selection dataset across repeated data thinning runs.}
\item{iv.hist.dt1}{Sample IV strength parameter in the inferential dataset based on signal-groups across repeated data thinning runs.}
\item{condF.dt1}{Conditional F-statistics in the selection dataset across repeated data thinning runs.}
\item{condF.dt2}{Conditional F-statistics in the inferential dataset based on signal-groups across repeated data thinning runs.}
}
\description{
Function to perform MVMR-PACS estimator for signal-group identification and variable selection, along with two-fold data thinning for post-selection inference.
}
