% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvmr.pacs.R
\name{mvmr.pacs}
\alias{mvmr.pacs}
\title{Function to perform MVMR-PACS estimator for signal-group identification and variable selection in two-sample multivariable Mendeliean randomization with highly-correlated exposures.}
\usage{
mvmr.pacs(
  beta.exposure,
  se.exposure,
  beta.outcome,
  se.outcome,
  P,
  type = 2,
  rr_cut_off = 0,
  fix.cor = FALSE,
  betawt = NULL,
  tau = c(0.5, 1, 2, 3),
  CV_fold = 5,
  n_times = 1,
  lambda.length = 20,
  eps = 1e-04,
  digit = 3,
  seed = NULL,
  lambda.1se = TRUE
)
}
\arguments{
\item{beta.exposure}{A data.frame or matrix. Each row contains the estimated marginal effect of a SNP on K exposures, usually obtained from a GWAS}

\item{se.exposure}{A data.frame or matrix of estimated standard errors of beta.exposure}

\item{beta.outcome}{A vector of the estimated marginal effect of a SNP on outcome, usually obtained from a GWAS}

\item{se.outcome}{A vector of estimated standard errors of beta.outcome}

\item{P}{A K-by-K matrix for the estimated shared correlation matrix between the effect of the genetic variants on each exposure, where K is the number of exposure.}

\item{type}{Weighting scheme in penalization: 1 = weights independent of correlation between beta.exposure; 2 = weights inversely related to correlation between beta.exposure; 3 = weights non-zero only if absolute correlation between beta.exposure greater than rr_cut_off; 4 = weights inversely related to correlation between beta.exposure only if absolute correlation between beta.exposure greater than rr_cut_off.}

\item{rr_cut_off}{Pre-specified cut-off used in different weighting schemes, specified by type argument.}

\item{fix.cor}{Whether or not recompute correlation of beta.exposure in cross validation. Default is FALSE.}

\item{betawt}{Initial beta coefficients. Default = NULL, meaning MVMR.dIVW with ridge penalty is used.}

\item{tau}{Tuning parameter range. Default = c(0.5, 1, 2, 3).}

\item{CV_fold}{The number of folds for cross validation based on multifold data thinning. Default = 5.}

\item{n_times}{The number of repeated cross validation. Default = 1.}

\item{lambda.length}{Length of candidate penalization parameter. Default = 20.}

\item{eps}{Tolerance level for ADMM.proj. Default = 1e-4.}

\item{digit}{Number of digits to keep when determining signal-groups. Default = 3, i.e., decimal precision to 0.001.}

\item{seed}{Seed for reproducibility. Default = NULL.}

\item{lambda.1se}{Whether or not to perform 1SE rules when selecting tuning parameters. Default = TRUE.}
}
\value{
A list with elements
\item{beta}{Estimated direct effects of each exposure on the outcome. Exposures with estimated effects of same magnitude (up to a prespecified decimal precision) belong to the same signal-group.}
\item{signal_group}{Identified signal-groups}
\item{iv_strength_parameter}{The minimum eigenvalue of the sample IV strength matrix, which quantifies the IV strength in the sample}
\item{lambda.index}{Index for the best penalization parameter lambda}
\item{lambda}{Value for the best penalization parameter lambda}
\item{tau.index}{Index for the best penalization parameter tau}
\item{tau}{Value for the best penalization parameter tau}
\item{beta.init}{Initial beta estimates}
}
\description{
Function to perform MVMR-PACS estimator for signal-group identification and variable selection in two-sample multivariable Mendeliean randomization with highly-correlated exposures.
}
