% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obtain_initial.R
\name{obtain_initial}
\alias{obtain_initial}
\title{Obtain initial beta coefficients using MVMR.dIVW with ridge penalty.}
\usage{
obtain_initial(
  beta.exposure,
  se.exposure,
  beta.outcome,
  se.outcome,
  iv_strength_parameter,
  P,
  CV_fold = 5,
  n_times = 1,
  epsilon = 1e-04,
  seed = NULL,
  lambda.1se = TRUE,
  lambda.length = 20
)
}
\arguments{
\item{beta.exposure}{A data.frame or matrix. Each row contains the estimated marginal effect of a SNP on K exposures, usually obtained from a GWAS}

\item{se.exposure}{A data.frame or matrix of estimated standard errors of beta.exposure}

\item{beta.outcome}{A vector of the estimated marginal effect of a SNP on outcome, usually obtained from a GWAS}

\item{se.outcome}{A vector of estimated standard errors of beta.outcome}

\item{iv_strength_parameter}{Sample IV strength parameter. Output from mvmr.srivw or mvmr.ivw estimators.}

\item{P}{A K-by-K matrix for the estimated shared correlation matrix between the effect of the genetic variants on each exposure, where K is the number of exposure.}

\item{CV_fold}{The number of folds for cross validation based on multifold data thinning. Default = 5.}

\item{n_times}{The number of repeated cross validation. Default = 1.}

\item{epsilon}{Tolerance level for ADMM.proj. Default = 1e-4.}

\item{seed}{Seed for reproducibility. Default = NULL.}

\item{lambda.1se}{Whether or not to perform 1SE rules when selecting tuning parameters. Default = TRUE.}

\item{lambda.length}{Length of candidate penalization parameter. Default = 20.}
}
\value{
\item{betawt}{Estimated direct effects of each exposure on the outcome.}
}
\description{
Obtain initial beta coefficients using MVMR.dIVW with ridge penalty.
}
